Fix #518: collapse duplicated JSON sidecar load/save boilerplate
Introduce JSONSidecarConfig protocol providing default load(from:)->Self?
and save(to:) implementations. ZoteroConfig, ExtractionConfig, and
AgentProvidersConfig now conform and drop their duplicated boilerplate;
each keeps only fileName + type-specific fields (loadOrSeed stays on
AgentProvidersConfig as genuinely type-specific). Public load/update APIs
unchanged.
